Add patch from robs@users.sf.net:
Adds support for FLAC input and output (if FLAC libs
are available).

Allows vorbis encoding quality to be specified.

Warns that with mp3, encoding quality cannot be
specified.

Fixes a bug in deemph that prevented its use with
ogg, flac, etc.

A couple of const-correctness fixes that were needed
for flac.

Documentation for the above.