Add x{m,c,re}alloc, and stop checking the return code of {m,c,realloc}, as it was not always done, when it was done it wasn't clear that the code could cope with the consequences, and sometimes it was done wrong, e.g.:
if (malloc(...) == NULL || malloc(...) == NULL || ...)
  st_fail(...)

which leaks if any malloc after the first fails.

Also, tidy up some code, e.g. code that was calling malloc and then
zeroing the result now calls xcalloc (with the exception of code
allocating floats or doubles: although on almost all architectures all
zero bits == 0.0, it's not universally true).

I've tagged the tree add-xmalloc in case anyone is horrified by this
change. In case you're worried about library-style use of libst,
don't: in sox, aborting on out of memory is never a problem, and in a
long-running application xmalloc can be redefined to call longjmp
instead of exit. In any case, as I said, the existing code was often
wrong.