Update the skel.c and skeleff.c files. Although I intend to remove them, I will be taking their didactic function from them, so if they look good first that's all to the good.
Make "make distcheck" build: this does a test "virgin" build,
effectively checking that SoX can be built from the files that are in
the distribution tarball; a versioned tarball is then built.

First cut at Lua scripting. Lua is a lightweight scripting language
(http://www.lua.org/) that has many advantages: it's pure ANSI C (the
same source is also valid ISO C++), so will build anywhere; it's
modest in size (12kLOC, ~100Kb of object code), interfaces easily to
C, and is simple. I've provided Lua support as an effect: the data is
marshalled into an array which is sent to Lua, a script given by the
user is run, and the data comes back. The idea is that it should be
much faster to hack up new filters in Lua than in C, and they
shouldn't run too slowly (in particular, other parts of SoX such as
FFT.c, can also be provided to Lua scripts).