Remove soxmix, since it's just sox -m now.
Detect pkg-config before trying to use it.

Remove some complication (ok, optimisation) from sox.c in order to
prepare for trying to simplify effects structure. In particular,
remove detection of null effects and of calls that don't write (the
old writing flag). The latter means that the entire file will be
processed even when only the header is read; this could be a problem,
but it should be done more simply. The former shouldn't be a problem
with performance-critical uses; if I'm wrong, I'm happy to look into
reinstating the optimisations in a more modular way.