Add support for ADPCM-encoded PRC files. To do this, move the ADPCM code from vox.c into adpcms.c, and make the functions public, adding st_adpcm_flush and st_adpcm_reset, because unlike vox, PRC files use frames, so need to reset the encoder while reading and writing.
Basically, you shouldn't write this format unless you have to, as
Psion Record seems to insist on frames of <=800 samples, which means
that the encoder is reset every tenth of a second. If you try this
with a pure tone, you'll find out why it's horrible. I've kept this
restriction for compatibility, though in theory it could be lifted
(with one frame per SoX buffer, i.e. by default once per 8192 bytes,
or roughly once a second, the results are fine).

Hence, the main use for this is for reading ADPCM-encoded PCM files
(the decoder copes with arbitrary frame lengths).

I owe a big debt of gratitude to David Smith (drsmith@iee.org) for his
sndcmp and rec2wav utilities, whose framing code I used, and whose
description of the file format I relied on (the code I used originates
from the Mathematisch Centrum, and was released under a free-use and
hence LGPL-compatible license).